test: adapt PUC-Rio Lua 5.1 test suite for LuaJIT
In some insignificant details LuaJIT's behaviour is not the same
as the behaviour of Lua 5.1.

Tests in the following files are disabled with corresponding comments:
* closure.lua
* constructs.lua
* db.lua
* errors.lua
* literals.lua
* main.lua
* math.lua
* nextvar.lua
* pm.lua
* strings.lua
* vararg.lua

jit.off() as prologue and jit.on() as epilogue is added to one test
chunk in gc.lua for stable (JIT unrelated) GC test results.

Part of tarantool/tarantool#5845
Part of tarantool/tarantool#4473