test: adapt Lua 5.1 test suite for Tarantool
In some insignificant details Tarantool's behaviour is not the same
as the behaviour of Lua 5.1.

This patch adds special LUAJIT_TEST_COMMAND variable to extend
`progname` from main.lua with ` -e require"strict".off()` to disable strict
for debug Tarantool build. Also, `progname` is not edged by '"' to be
able run other test suites.

Tests in the following files are disabled with corresponding comments:
* gc.lua
* main.lua
* pm.lua

Tests in the following files are adapted for testing with Tarantool's
out-of-source build:
* all.lua
* verybig.lua

Part of tarantool/tarantool#5845
Part of tarantool/tarantool#4473