test: disable PUC-Rio test for tailcall info
LuaJIT does not provide information for tailcalls in `debug.getinfo()`,
unlike Lua does. This missed feature is described in
https://luajit.org/status.html.

This patch disables tests for tailcall checks and `getfenv()` checks,
because tailcalls do not provide an additional call frame for LuaJIT
and level number given to `getfenv()` should be changed.

Relates to tarantool/tarantool#5702
Relates to tarantool/tarantool#5703
Part of tarantool/tarantool#5845
Part of tarantool/tarantool#4473

Reviewed-by: Sergey Ostanevich <sergos@tarantool.org>
Reviewed-by: Igor Munkin <imun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>