test: disable PUC-Rio test for getfenv in tailcall
Tail calls are not taken into account for call-level counting. So,
`getfenv()` behaviour in LuaJIT is different from Lua 5.1 in tailcalls.
LuaJIT doesn't create a new frame for tailcalls, just uses the caller's.

This patch disables test for the return result of `getfenv()` tailcall.
Default value (equals to 1) of `getfenv()` function's argument (function
level) is invalid for this tailcall -- LuaJIT can't provide necessary
debug information for the frame.

Relates to tarantool/tarantool#5713
Part of tarantool/tarantool#5845
Part of tarantool/tarantool#4473

Reviewed-by: Sergey Ostanevich <sergos@tarantool.org>
Reviewed-by: Igor Munkin <imun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>