test: fix directory detection in lua-Harness suite
A test <314-regex.t> uses `arg[0]:find'314'` to determine the name of
the directory where rx_* files are located. This leads to the test
failure, when lua-Harness suite runs in a directory containing "314" in
its name, because the found path doesn't contain the required files.

This patch fixes directory name detection.

Follows up tarantool/tarantool#5844

Reviewed-by: Igor Munkin <imun@tarantool.org>
Reviewed-by: Sergey Ostanevich <sergos@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>