sysprof: fix segfaults under load
There was a rare bug related to unconsistency of L->base when
profiling occurs, which was reproduced only under load:

In function <dump_lfunc> function <lj_debug_frameline>, which is
not resilient to L->base unconsistency, was called.

The solution is simple: we just got rid of that function as we don't
need it really: it was used to get line where the function was called,
but we actually need the line where the function was defined.